+++ client/CHANGELOG.md
@@ -1,3 +1,46 @@
+# client [1.33.0-rc.4](https://github.com/fhswf/appointme/compare/client@1.33.0-rc.3...client@1.33.0-rc.4) (2026-01-21)
+
+
+### Bug Fixes
+
+* make privacy the main item on the legal page ([0713a35](https://github.com/fhswf/appointme/commit/0713a35f777f6d05696ef0aaf86e98efccefbcf5))
+* update legal information on the landing page ([7854f2d](https://github.com/fhswf/appointme/commit/7854f2dbd2b31c6a0461a41cef5331619f967e90))
+
+
+
+
+
+### Dependencies
+
+* **common:** upgraded to 1.15.1-rc.1
+
+# client [1.33.0-rc.3](https://github.com/fhswf/appointme/compare/client@1.33.0-rc.2...client@1.33.0-rc.3) (2026-01-21)
+
+
+### Features
+
+* import/export user settings ([72ed714](https://github.com/fhswf/appointme/commit/72ed7142b51026d7f03081b4c6a403507f9e02ac))
+
+# client [1.33.0-rc.2](https://github.com/fhswf/appointme/compare/client@1.33.0-rc.1...client@1.33.0-rc.2) (2026-01-20)
+
+
+### Bug Fixes
+
+* authentication handling ([e25b793](https://github.com/fhswf/appointme/commit/e25b7938be4225be4bf3851de036fe3b79da8c4a))
+* authentication handling ([264b701](https://github.com/fhswf/appointme/commit/264b701d3aa38d680879ea214e77f208f05f958b))
+
+
+### Features
+
+* separate access token from "transient" lti token ([da9e2e7](https://github.com/fhswf/appointme/commit/da9e2e7a6362ab65fd2c07b6cc12c685fc349d1e))
+
+# client [1.33.0-rc.1](https://github.com/fhswf/appointme/compare/client@1.32.0...client@1.33.0-rc.1) (2026-01-20)
+
+
+### Features
+
+* separate access token from "transient" lti token ([a8d1482](https://github.com/fhswf/appointme/commit/a8d1482b1429a3b8b6a0710183d2f24f67c6798a))
+
 # client [1.32.0](https://github.com/fhswf/appointme/compare/client@1.31.0...client@1.32.0) (2026-01-20)